refactor: rename summarizer_settings → summarizerSettings
