+++ gitbutler-ui/src/lib/backend/summarizerSettings.ts
